questionnaire/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
djangorestframework==3.13.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
